wip refactor: column
