+++ console/packages/starwhale-ui/src/GridTable/GridTable.tsx
@@ -80,7 +80,6 @@ function GridTable({
     useEffect(() => {
         if (!store.isInit && columns.length > 0) {
             api.setState({
-                isInit: true,
                 currentView: {
                     name: '',
                     filters: [],
@@ -101,41 +100,44 @@ function GridTable({
                 className={cn(styles.table, styles.tablePinnable, compareable ? styles.tableCompareable : undefined)}
                 ref={wrapperRef}
             >
-                <StatefulDataTable
-                    store={store}
-                    useStore={api}
-                    resizableColumnWidths
-                    initialFilters={$filters}
-                    searchable={searchable}
-                    filterable={filterable}
-                    queryable={queryable}
-                    columnable={columnable}
-                    compareable={compareable}
-                    viewable={viewable}
-                    loading={!!isLoading}
-                    rowActions={rowActions}
-                    columns={columns}
-                    rows={$rows}
-                    // @ts-ignore
-                    loadingMessage={() =>
-                        (
-                            <Skeleton
-                                overrides={{
-                                    Root: {
-                                        style: {
-                                            paddingTop: '10px',
+                {isLoading && <BusyPlaceholder />}
+                {!isLoading && (
+                    <StatefulDataTable
+                        store={store}
+                        useStore={api}
+                        resizableColumnWidths
+                        initialFilters={$filters}
+                        searchable={searchable}
+                        filterable={filterable}
+                        queryable={queryable}
+                        columnable={columnable}
+                        compareable={compareable}
+                        viewable={viewable}
+                        loading={!!isLoading}
+                        rowActions={rowActions}
+                        columns={columns}
+                        rows={$rows}
+                        // @ts-ignore
+                        loadingMessage={() =>
+                            (
+                                <Skeleton
+                                    overrides={{
+                                        Root: {
+                                            style: {
+                                                paddingTop: '10px',
+                                            },
                                         },
-                                    },
-                                }}
-                                rows={10}
-                                width='100%'
-                                animation
-                            />
-                        ) as any
-                    }
-                    // @ts-ignore
-                    emptyMessage={() => (<BusyPlaceholder type='notfound' />) as any}
-                />
+                                    }}
+                                    rows={10}
+                                    width='100%'
+                                    animation
+                                />
+                            ) as any
+                        }
+                        // @ts-ignore
+                        emptyMessage={() => (<BusyPlaceholder type='notfound' />) as any}
+                    />
+                )}
             </div>
             <Pagination {...paginationProps} />
         </>
